Switch to TransitionGroup rather than CSSTransitionGroup.
Means we actually get the lifecycle events for firing off transition events,
which in turn means that we can use the TransitionEnder class to do a better 
job to figuring out when transitions actually fire/end/etc.

The implementation in AnimationItem should match the general set up in the
CSSTransitionGroup implementation.