Use a second Thing search to provide non-matching fallback results
This uses the same _filters_ as the main search (e.g. exclude_mature, exclude_sensitive, etc.), but negates the full-text search terms so that there is no overlap in results.

Add an `exclude_fallbacks` param to the Api::SearchController to determine whether to include these fallback results. The default is to include them.